Add Chaotic Noise support to PNG8

+++ src/main/java/com/github/tommyettinger/anim8/AnimatedGif.java
@@ -408,7 +408,7 @@ public class AnimatedGif implements AnimationWriter, Dithered {
             break;
             case BLUE_NOISE: { 
                 float adj, strength = palette.ditherStrength;
-                byte bn;
+                int bn;
                 for (int y = 0, i = 0; y < height && i < nPix; y++) {
                     for (int px = 0; px < width & i < nPix; px++) {
                         color = image.getPixel(px, flipped + flipDir * y) & 0xF8F8F880;
